move request to common
