+++ packages/dd-trace/src/exporters/agent/writer.js
@@ -1,6 +1,6 @@
 'use strict'
 
-const request = require('./request')
+const request = require('../common/request')
 const { startupLog } = require('../../startup-log')
 const metrics = require('../../metrics')
 const log = require('../../log')
@@ -99,7 +99,7 @@ function makeRequest (version, data, count, url, lookup, needsStartupLog, cb) {
 
   log.debug(() => `Request to the agent: ${JSON.stringify(options)}`)
 
-  request(Object.assign({ data }, options), (err, res, status) => {
+  request(data, options, true, (err, res, status) => {
     if (needsStartupLog) {
       // Note that logging will only happen once, regardless of how many times this is called.
       startupLog({
