+++ packages/dd-trace/test/exporters/common/docker.spec.js
@@ -11,7 +11,7 @@ describe('docker', () => {
   })
 
   it('should return an empty ID when the cgroup cannot be read', () => {
-    docker = proxyquire('../src/exporters/agent/docker', { fs })
+    docker = proxyquire('../src/exporters/common/docker', { fs })
 
     expect(docker.id()).to.be.undefined
   })
@@ -22,7 +22,7 @@ describe('docker', () => {
     ].join('\n')
 
     fs.readFileSync.withArgs('/proc/self/cgroup').returns(Buffer.from(cgroup))
-    docker = proxyquire('../src/exporters/agent/docker', { fs })
+    docker = proxyquire('../src/exporters/common/docker', { fs })
 
     expect(docker.id()).to.equal('34dc0b5e626f2c5c4c5170e34b10e7654ce36f0fcd532739f4445baabea03376')
   })
@@ -33,7 +33,7 @@ describe('docker', () => {
     ].join('\n')
 
     fs.readFileSync.withArgs('/proc/self/cgroup').returns(Buffer.from(cgroup))
-    docker = proxyquire('../src/exporters/agent/docker', { fs })
+    docker = proxyquire('../src/exporters/common/docker', { fs })
 
     expect(docker.id()).to.equal('34dc0b5e-626f-2c5c-4c51-70e34b10e765')
   })
@@ -44,7 +44,7 @@ describe('docker', () => {
     ].join('\n')
 
     fs.readFileSync.withArgs('/proc/self/cgroup').returns(Buffer.from(cgroup))
-    docker = proxyquire('../src/exporters/agent/docker', { fs })
+    docker = proxyquire('../src/exporters/common/docker', { fs })
 
     expect(docker.id()).to.equal('34dc0b5e626f2c5c4c5170e34b10e765-1234567890')
   })
@@ -55,7 +55,7 @@ describe('docker', () => {
     ].join('\n')
 
     fs.readFileSync.withArgs('/proc/self/cgroup').returns(Buffer.from(cgroup))
-    docker = proxyquire('../src/exporters/agent/docker', { fs })
+    docker = proxyquire('../src/exporters/common/docker', { fs })
 
     expect(docker.id()).to.equal('7b8952daecf4c0e44bbcefe1b5c5ebc7b4839d4eefeccefe694709d3809b6199')
   })
@@ -68,7 +68,7 @@ describe('docker', () => {
     ].join('\n')
 
     fs.readFileSync.withArgs('/proc/self/cgroup').returns(Buffer.from(cgroup))
-    docker = proxyquire('../src/exporters/agent/docker', { fs })
+    docker = proxyquire('../src/exporters/common/docker', { fs })
 
     expect(docker.id()).to.equal('34dc0b5e626f2c5c4c5170e34b10e7654ce36f0fcd532739f4445baabea03376')
   })
@@ -79,7 +79,7 @@ describe('docker', () => {
     ].join('\n')
 
     fs.readFileSync.withArgs('/proc/self/cgroup').returns(Buffer.from(cgroup))
-    docker = proxyquire('../src/exporters/agent/docker', { fs })
+    docker = proxyquire('../src/exporters/common/docker', { fs })
 
     expect(docker.id()).to.equal('34dc0b5e626f2c5c4c5170e34b10e7654ce36f0fcd532739f4445baabea03376')
   })
